Rename services layer to usecases
